siatex

+++ resources/views/attendance/entry.blade.php
@@ -87,30 +87,32 @@
                                 <tr>
 
                                     <th scope="col"
-                                        class="p-4 w-2/12 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                        class="p-4 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
                                         Name
                                     </th>
                                     <th scope="col"
-                                        class="p-4 w-2/12 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                        class="p-4  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
                                         Status
                                     </th>
                                     <th scope="col"
-                                        class="p-4 w-2/12 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                        class="p-4  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
                                         Entry Time
                                     </th>
                                     <th scope="col"
-                                        class="p-4 w-2/12 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                        class="p-4  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
                                         Entry By
                                     </th>
                                     <th scope="col"
-                                        class="p-4 w-2/12 text-xs font-medium text-left text-gray-500 uppercase dark:text-gray-400">
+                                        class="p-4  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
                                         Comment
                                     </th>
+                                    @if (auth()->user()->role->name === 'admin')
+                                        <th scope="col"
+                                            class="p-4  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
+                                            Actions
+                                        </th>
+                                    @endif
 
-                                    <th scope="col"
-                                        class="p-4 w-2/12  text-xs font-medium text-center text-gray-500 uppercase dark:text-gray-400">
-                                        Actions
-                                    </th>
                                 </tr>
                             </thead>
                             <tbody class="bg-white divide-y divide-gray-200 dark:bg-gray-800 dark:divide-gray-700">
@@ -123,7 +125,7 @@
 
                                         </td>
                                         <td
-                                            class="p-4 text-base font-medium text-gray-900 whitespace-normal dark:text-white">
+                                            class="p-4 text-center text-base font-medium text-gray-900 whitespace-normal dark:text-white">
                                             @php
                                                 $checkInDateTime = \Carbon\Carbon::parse($item->check_in);
                                                 $expectedStartTime = \Carbon\Carbon::parse('09:00')->setDate($checkInDateTime->year, $checkInDateTime->month, $checkInDateTime->day);
@@ -142,52 +144,54 @@
 
 
                                         <td
-                                            class="p-4  text-base font-medium text-gray-900 whitespace-normal dark:text-white">
+                                            class="p-4 text-center  text-base font-medium text-gray-900 whitespace-normal dark:text-white">
                                             {{ $checkInDateTime->format('F d, Y h:i A') }}
 
                                         </td>
 
                                         <td
-                                            class="p-4  uppercase text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
+                                            class="p-4 text-center  uppercase text-base font-medium text-gray-900 whitespace-nowrap dark:text-white">
                                             {{ $item->user->name }}
                                         </td>
                                         <td
-                                            class="p-4  text-base font-medium text-gray-900 {{ $item->comment ? 'min-w-[300px] ' : '' }} whitespace-normal dark:text-white">
+                                            class="p-4 text-center text-base font-medium text-gray-900 {{ $item->comment ? 'min-w-[300px] ' : '' }} whitespace-normal dark:text-white">
                                             <span>{{ $item->comment }}</span>
                                         </td>
 
-
-                                        <td class="p-4 text-right space-x-2 whitespace-nowrap">
-                                            <button onclick="showUpdatePrevAttendanceData( {{ $item->id }})"
-                                                type="button" data-modal-target="attendance-update-modal"
-                                                data-modal-toggle="attendance-update-modal"
-                                                class="inline-flex items-center px-4 py-2 text-sm font-medium text-center text-white rounded-lg bg-lime-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800">
-                                                <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
-                                                    xmlns="http://www.w3.org/2000/svg">
-                                                    <path
-                                                        d="M17.414 2.586a2 2 0 00-2.828 0L7 10.172V13h2.828l7.586-7.586a2 2 0 000-2.828z">
-                                                    </path>
-                                                    <path fill-rule="evenodd"
-                                                        d="M2 6a2 2 0 012-2h4a1 1 0 010 2H4v10h10v-4a1 1 0 112 0v4a2 2 0 01-2 2H4a2 2 0 01-2-2V6z"
-                                                        clip-rule="evenodd"></path>
-                                                </svg>
-                                                Edit
-                                            </button>
-                                            <button type="button" id=""
-                                                onclick="deleteAttendance({{ $item->id }})"
-                                                class="inline-flex
+                                        @if (auth()->user()->role->name === 'admin')
+                                            <td class="p-4 text-right space-x-2 whitespace-nowrap">
+                                                <button onclick="showUpdatePrevAttendanceData( {{ $item->id }})"
+                                                    type="button" data-modal-target="attendance-update-modal"
+                                                    data-modal-toggle="attendance-update-modal"
+                                                    class="inline-flex items-center px-4 py-2 text-sm font-medium text-center text-white rounded-lg bg-lime-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800">
+                                                    <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
+                                                        xmlns="http://www.w3.org/2000/svg">
+                                                        <path
+                                                            d="M17.414 2.586a2 2 0 00-2.828 0L7 10.172V13h2.828l7.586-7.586a2 2 0 000-2.828z">
+                                                        </path>
+                                                        <path fill-rule="evenodd"
+                                                            d="M2 6a2 2 0 012-2h4a1 1 0 010 2H4v10h10v-4a1 1 0 112 0v4a2 2 0 01-2 2H4a2 2 0 01-2-2V6z"
+                                                            clip-rule="evenodd"></path>
+                                                    </svg>
+                                                    Edit
+                                                </button>
+                                                <button type="button" id=""
+                                                    onclick="deleteAttendance({{ $item->id }})"
+                                                    class="inline-flex
                                             items-center px-3 py-2 text-sm font-medium text-center text-white bg-red-600
                                             rounded-lg hover:bg-red-800 focus:ring-4 focus:ring-red-300
                                             dark:focus:ring-red-900">
-                                                <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
-                                                    xmlns="http://www.w3.org/2000/svg">
-                                                    <path fill-rule="evenodd"
-                                                        d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z"
-                                                        clip-rule="evenodd"></path>
-                                                </svg>
-                                                Delete
-                                            </button>
-                                        </td>
+                                                    <svg class="w-4 h-4 mr-2" fill="currentColor" viewBox="0 0 20 20"
+                                                        xmlns="http://www.w3.org/2000/svg">
+                                                        <path fill-rule="evenodd"
+                                                            d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z"
+                                                            clip-rule="evenodd"></path>
+                                                    </svg>
+                                                    Delete
+                                                </button>
+                                            </td>
+                                        @endif
+
                                     </tr>
                                 @endforeach
                             </tbody>
